refactor: reorganize utils to prompting with clearer separation of concerns
BREAKING CHANGE: This is a major refactoring that improves code organization
and naming clarity. Applications using the old interface must update their code.

Changes:
- Rename directory from src/utils/ to src/prompting/ for clarity
- Split promptBuilder.ts into builder.ts and parser.ts for better separation
- Move extractRandomVariables to content.ts where it logically belongs
- Rename parseMessagesFromTemplate to buildMessagesFromTemplate
- Update all imports and documentation references

Migration required:
- Update imports from 'genai-lite/utils' to 'genai-lite/prompting'
- Rename parseMessagesFromTemplate calls to buildMessagesFromTemplate
- See MIGRATION_UTILS_TO_PROMPTING.md for detailed migration guide

Benefits:
- Clearer directory name immediately indicates purpose
- Better file organization by function (template, content, builder, parser)
- Logical workflow alignment: prepare content → build messages → parse responses
- Improved function naming that better reflects actual purpose
