+++ .summary_long.md
@@ -251,5 +251,5 @@ import { LLMService, fromEnvironment } from 'genai-lite';
   - .github/ - GitHub-specific configuration
   - src/llm/ - Core LLM service implementation
   - src/providers/ - API key provider implementations
-  - src/utils/ - Utility functions
+  - src/prompting/ - Prompt engineering utilities
 - **Key Consumers**: Any Node.js application requiring unified AI provider access
