+++ README.md
@@ -296,14 +296,14 @@ import type {
 
 ## Utilities
 
-genai-lite includes useful utilities for working with LLMs, available through the `genai-lite/utils` subpath:
+genai-lite includes useful utilities for working with LLMs, available through the `genai-lite/prompting` subpath:
 
 ### Token Counting
 
 Count the number of tokens in a string using OpenAI's tiktoken library:
 
 ```typescript
-import { countTokens } from 'genai-lite/utils';
+import { countTokens } from 'genai-lite/prompting';
 
 const text = 'Hello, this is a sample text for token counting.';
 const tokenCount = countTokens(text); // Uses gpt-4 tokenizer by default
@@ -320,7 +320,7 @@ const gpt35Tokens = countTokens(text, 'gpt-3.5-turbo');
 Generate intelligent previews of large text blocks that preserve context:
 
 ```typescript
-import { getSmartPreview } from 'genai-lite/utils';
+import { getSmartPreview } from 'genai-lite/prompting';
 
 const largeCodeFile = `
 function calculateTotal(items) {
@@ -359,7 +359,7 @@ Combine these utilities to build prompts that fit within model context windows:
 
 ```typescript
 import { LLMService, fromEnvironment } from 'genai-lite';
-import { countTokens, getSmartPreview } from 'genai-lite/utils';
+import { countTokens, getSmartPreview } from 'genai-lite/prompting';
 
 const llm = new LLMService(fromEnvironment);
 
@@ -394,7 +394,7 @@ const response = await llm.sendMessage({
 Generate dynamic prompts and content using the built-in template engine that supports variable substitution and conditional logic:
 
 ```typescript
-import { renderTemplate } from 'genai-lite/utils';
+import { renderTemplate } from 'genai-lite/prompting';
 
 // Simple variable substitution
 const greeting = renderTemplate('Hello, {{ name }}!', { name: 'World' });
@@ -432,7 +432,7 @@ const result = renderTemplate(complexTemplate, {
   expertise: 'TypeScript, React, Node.js',
   task: 'Review the code for best practices',
   hasFiles: true,
-  fileList: '- src/index.ts\n- src/utils.ts',
+  fileList: '- src/index.ts\n- src/prompting/template.ts',
   requiresOutput: false
 });
 ```
@@ -450,7 +450,7 @@ Combine the template engine with other utilities for powerful prompt generation:
 
 ```typescript
 import { LLMService, fromEnvironment } from 'genai-lite';
-import { renderTemplate, countTokens } from 'genai-lite/utils';
+import { renderTemplate, countTokens } from 'genai-lite/prompting';
 
 const llm = new LLMService(fromEnvironment);
 
@@ -500,7 +500,7 @@ genai-lite provides powerful utilities for building and parsing structured promp
 Convert template strings with role tags into LLM message arrays:
 
 ```typescript
-import { parseMessagesFromTemplate } from 'genai-lite/utils';
+import { buildMessagesFromTemplate } from 'genai-lite/prompting';
 
 const template = `
 <SYSTEM>You are a helpful assistant specialized in {{expertise}}.</SYSTEM>
@@ -509,7 +509,7 @@ const template = `
 <USER>Can you provide more details?</USER>
 `;
 
-const messages = parseMessagesFromTemplate(template, {
+const messages = buildMessagesFromTemplate(template, {
   expertise: 'TypeScript and React',
   task: 'building a custom hook'
 });
@@ -528,7 +528,7 @@ const messages = parseMessagesFromTemplate(template, {
 Implement few-shot prompting by extracting and shuffling examples:
 
 ```typescript
-import { extractRandomVariables, renderTemplate } from 'genai-lite/utils';
+import { extractRandomVariables, renderTemplate } from 'genai-lite/prompting';
 
 // Define examples in your template
 const examplesTemplate = `
@@ -570,7 +570,7 @@ const prompt = renderTemplate(promptTemplate, {
 Extract structured data from LLM responses using custom tags:
 
 ```typescript
-import { parseStructuredContent } from 'genai-lite/utils';
+import { parseStructuredContent } from 'genai-lite/prompting';
 
 // Example LLM response with structured output
 const llmResponse = `
