+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "genai-lite",
-  "version": "0.1.4",
+  "version": "0.2.0",
   "description": "A lightweight, portable toolkit for interacting with various Generative AI APIs.",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
@@ -10,10 +10,10 @@
       "require": "./dist/index.js",
       "types": "./dist/index.d.ts"
     },
-    "./utils": {
-      "import": "./dist/utils/index.js",
-      "require": "./dist/utils/index.js",
-      "types": "./dist/utils/index.d.ts"
+    "./prompting": {
+      "import": "./dist/prompting/index.js",
+      "require": "./dist/prompting/index.js",
+      "types": "./dist/prompting/index.d.ts"
     }
   },
   "author": "Luigi Acerbi <luigi.acerbi@gmail.com>",
