+++ src/.summary_long.md
@@ -151,5 +151,5 @@ console.log(response);
 - **Related Directories**:
   - `src/llm/` - Core LLM service implementation and configurations
   - `src/providers/` - API key provider implementations
-  - `src/utils/` - Utility functions and helpers
+  - `src/prompting/` - Prompt engineering utilities (template rendering, content preparation, message building, response parsing)
 - **Key Consumers**: All applications and services that depend on genai-lite import from this directory
