+++ src/.summary_short.md
@@ -13,4 +13,4 @@ Purpose: Contains the main source code for the genai-lite library, providing the
 - `types/`: Contains type definitions including the ModelPreset interface for configurable presets.
 - `llm/`: Core LLM service implementation with support for configurable presets.
 - `providers/`: Built-in API key provider implementations.
-- `utils/`: Utility functions for prompt processing.
+- `prompting/`: Sophisticated utilities for prompt engineering including template rendering, content preparation, message building, and response parsing.
