+++ src/llm/.summary_long.md
@@ -370,5 +370,5 @@ const response = await service.sendMessage({
 - **Related Directories**:
   - src/llm/clients/ - Provider-specific adapter implementations
   - src/providers/ - API key provider implementations
-  - src/utils/ - Utility functions
+  - src/prompting/ - Prompt engineering utilities
 - **Key Consumers**: All applications using genai-lite interact primarily through this directory's exports
