+++ src/prompting/.summary_short.md
@@ -1,0 +1,15 @@
+# Directory: src/prompting/
+
+Purpose: Contains sophisticated utilities for prompt engineering, including template rendering, content preparation, structured message building, and response parsing for LLM interactions.
+
+## Files:
+
+- `index.ts`: Central export point for all prompting utilities, providing a clean API for template rendering, content preparation, message building, and response parsing.
+- `template.ts`: Implements a sophisticated micro-templating engine with variable substitution, conditional ternary logic, and intelligent newline handling.
+- `template.test.ts`: Comprehensive test suite for the template engine, validating variable substitution, ternary expressions, nested templates, and newline trimming behavior.
+- `content.ts`: Provides content preparation utilities including token counting using OpenAI's tiktoken, smart content previews, and random variable extraction for few-shot learning.
+- `content.test.ts`: Tests content preparation utilities including token counting accuracy across models, smart preview truncation logic, and random variable extraction.
+- `builder.ts`: Implements message building functionality that converts templates with role tags (SYSTEM, USER, ASSISTANT) into structured LLM message arrays.
+- `builder.test.ts`: Tests message building from templates, variable substitution within role tags, and proper message ordering.
+- `parser.ts`: Provides utilities for parsing structured content from LLM responses using XML-style tags, supporting both closed and unclosed tag formats.
+- `parser.test.ts`: Tests structured content parsing functionality, including tag extraction, handling of missing tags, and multiline content parsing.
