+++ src/prompting/index.ts
@@ -1,0 +1,21 @@
+/**
+ * Prompting utilities for advanced LLM prompt engineering
+ * 
+ * This module provides a complete toolkit for prompt engineering workflows:
+ * - Template rendering with variable substitution and conditionals
+ * - Content preparation and analysis (token counting, smart previews)
+ * - Prompt building from templates to structured messages
+ * - Response parsing for extracting structured data from LLM outputs
+ */
+
+// Template rendering
+export { renderTemplate } from './template';
+
+// Content preparation utilities
+export { countTokens, getSmartPreview, extractRandomVariables } from './content';
+
+// Prompt building
+export { buildMessagesFromTemplate } from './builder';
+
+// Response parsing
+export { parseStructuredContent } from './parser';
