--- src/utils/.summary_short.md
@@ -1,13 +1,0 @@
-# Directory: src/utils/
-
-Purpose: Contains utility functions for prompt processing, including token counting, text preview generation, template rendering, and advanced prompt building utilities for LLM interactions.
-
-## Files:
-
-- `index.ts`: Exports utility functions for prompt processing, token counting, template rendering, and prompt building.
-- `prompt.ts`: Provides utility functions for counting tokens in text using OpenAI's tiktoken library and generating smart content previews with configurable line limits.
-- `prompt.test.ts`: Tests the prompt utilities module's token counting and smart text preview functions to ensure accurate tokenization across different models and intelligent content truncation.
-- `promptBuilder.ts`: Implements advanced prompt engineering utilities for parsing structured prompts with role tags, extracting random variables for few-shot learning, and parsing structured content from LLM responses.
-- `promptBuilder.test.ts`: Comprehensive test suite for prompt builder utilities, validating message parsing, random variable extraction, and structured content parsing functionality.
-- `templateEngine.ts`: Implements a sophisticated micro-templating engine that supports variable substitution, conditional ternary logic with backtick-delimited strings, and intelligent newline handling.
-- `templateEngine.test.ts`: Comprehensive test suite for the template engine, validating variable substitution, ternary expressions, nested templates, and newline trimming behavior.
